pyproject.toml
[project]
name = "wol-proxy"
version = "0.1.0"
description = "Wake on LAN proxy for my homelab"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "wakeonlan==3.1.0",
    "python-multipart==0.0.20",
    "jinja2==3.1.6",
    "httpx>=0.28.1",
    "asyncio>=3.4.3",
    "python-dotenv>=1.1.0",
]
